+++ src/components/customForm/useGetCustomFormElements.ts
@@ -7,6 +7,8 @@ import {
     useBoundResourceQueryFn,
     DEFAULT_CATEGORY_OPTION_COMBO,
 } from '../../lib'
+import { ProgramStateSectionDataElements } from '../../pages/programsWip/form/programStage/programStageSection/ProgramStageSectionFormContents'
+import { ElementTypes } from './CustomFormElementsSelector'
 
 type FlagItemResponse = { name: string; key: string; path: string }
 
@@ -65,7 +67,11 @@ type ProgramAttributeData = {
 export const useProgramsCustomFormElements = () => {
     const programId = useParams().id
     const queryFn = useBoundResourceQueryFn()
-    const { data: attributesData, isLoading } = useQuery({
+    const {
+        data: attributesData,
+        refetch,
+        isLoading,
+    } = useQuery({
         queryFn: queryFn<ProgramAttributeData>,
         queryKey: [
             {
@@ -81,7 +87,7 @@ export const useProgramsCustomFormElements = () => {
     })
 
     const elementTypes = useMemo(() => {
-        const attributes = attributesData?.programTrackedEntityAttributes.map(
+        const attributes = attributesData?.programTrackedEntityAttributes?.map(
             (ptea) => ptea.trackedEntityAttribute
         )
         const programElements: ElementItem[] = [
@@ -99,10 +105,10 @@ export const useProgramsCustomFormElements = () => {
                 elements: programElements,
                 type: 'program',
             },
-        ]
+        ] as ElementTypes
     }, [attributesData])
 
-    return { loading: isLoading, elementTypes }
+    return { loading: isLoading, refetch, elementTypes }
 }
 
 export const useDataSetCustomFormElements = () => {
@@ -112,7 +118,11 @@ export const useDataSetCustomFormElements = () => {
     const indicators = indicatorsInput.value
 
     const queryFn = useBoundResourceQueryFn()
-    const { data: deoData, isLoading } = useQuery({
+    const {
+        data: deoData,
+        isLoading,
+        refetch,
+    } = useQuery({
         queryFn: queryFn<DEOData>,
         queryKey: [
             {
@@ -178,7 +188,7 @@ export const useDataSetCustomFormElements = () => {
 
     // get flags
     const flagsQuery = useGetFlags()
-    const flags = flagsQuery.data || []
+    const flags = useMemo(() => flagsQuery.data || [], [flagsQuery])
 
     const elementTypes = useMemo(
         () => [
@@ -198,5 +208,57 @@ export const useDataSetCustomFormElements = () => {
         [dataElements, totals, indicators, flags]
     )
 
-    return { loading: flagsQuery.isLoading ?? isLoading, elementTypes }
+    return { loading: flagsQuery.isLoading ?? isLoading, refetch, elementTypes }
+}
+
+export const useProgramsStageSectionCustomFormElements = (stageId: string) => {
+    const queryFn = useBoundResourceQueryFn()
+    const {
+        data: deData,
+        isLoading,
+        refetch,
+    } = useQuery({
+        queryFn: queryFn<ProgramStateSectionDataElements>,
+        queryKey: [
+            {
+                resource: 'programStages',
+                id: stageId,
+                params: {
+                    fields: [
+                        'programStageDataElements[id,dataElement[id,displayName,categoryCombo[id]]]',
+                    ].concat(),
+                },
+            },
+        ] as const,
+    })
+
+    const dataElements = useMemo(() => {
+        if (!deData) {
+            return []
+        }
+        return deData.programStageDataElements
+            ?.map((de) => ({
+                id: `${stageId}-${de.dataElement.id}`,
+                displayName: de.dataElement.displayName,
+            }))
+            .sort((de1, de2) => de1.displayName.localeCompare(de2.displayName))
+    }, [deData, stageId])
+
+    // get flags
+    const flagsQuery = useGetFlags()
+    const flags = useMemo(() => flagsQuery.data || [], [flagsQuery])
+
+    const elementTypes = useMemo(
+        () => [
+            {
+                name: i18n.t('Data elements'),
+                elements: dataElements,
+                type: 'dataElement',
+            },
+            { name: i18n.t('Flags'), elements: flags, type: 'flag' },
+        ],
+        [dataElements, flags]
+    )
+
+    return { loading: flagsQuery.isLoading ?? isLoading, refetch, elementTypes }
 }